+++ Changelog.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-##[Unreleased]
+## 3.18.1
 ### Fixed
  - FOR-2158: Fixed lazyLoad issue with multiple select when viewing value
+ - FOR-2165: Fixed infinite loop with calculated value for select.
 
 ## 3.18.0
 ### Added